Add regmks and reglen
Record region markers to greatly reduce the amount of metadata update. As if
there is no editing within the region the relative distance of the two region
markers will remain the same. Unlike the old position indicator, any
modification before the marker will result in a position change and thus a need
for metadata update.